requirements.txt
# Church Management System
# Python dependencies for the application

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.12

# Forms and Validation
WTForms==3.1.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
